// src/api/models/CreateCategoryDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateCategoryDto = {
  /**
   * 分类名称
   */
  name: string;
  /**
   * 分类描述
   */
  description?: string;
  /**
   * 分类图标
   */
  icon?: string;
  /**
   * 分类排序
   */
  order?: number;
  /**
   * 分类是否激活
   */
  isActive?: boolean;
};